requirements.txt
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
requests>=2.31.0
scikit-learn>=1.3.0
